feat(actix-web): websocket example for actix-web
The example is similar to the one I found in the axum example although not really the same. This one adds an example for echoing messages on top of the ability to check how many users are connected at a given time and the status of the shuttle api.

While checking the axum example I saw some issues. I'll prepare a PR for that, too.