ci: Create ci.sh, run fmt and clippy fix
Run `cargo clippy --fix` and `cargo fmt` on all directories so that
less crates fail the ci.

Currently, the next/hello-world crate fails cargo clippy,
because cargo-clippy assumes the arch to be the same as
the device running it, while shuttle-next assumes that it is
being run on wasi.

This commit doesn't include the actual action yml file, which
will get added in the next commit